// src/contexts/TaskContext/TaskContextProvider.tsx
import { useEffect, useReducer, useRef } from "react";
import { initialTaskState } from "./initialTaskState";
import { TaskContext } from "./TaskContext";
import { taskReducer } from "./taskReducer";
import { TimerWorkerManager } from "../../workers/TimerWorkManager";
import { TaskActionsTypes } from "./taskActions";
import { loadBeep } from "../../utils/loadBeep";
import type { TaskStateModel } from "../../models/TaskStateModel";

type TaskContextProviderProps = {
    children: React.ReactNode
};

export function TaskContextProvider({ children }: TaskContextProviderProps) {
    const [state, dispatch] = useReducer(taskReducer, initialTaskState, () => {
        const storageState = localStorage.getItem('state')

        if (!storageState) return initialTaskState

        const parsedStorageState = JSON.parse(storageState) as TaskStateModel

        return {
            ...parsedStorageState,
            activeTask: null,
            secondsRemaining: 0,
            formattedSecondsRemaining: '00:00',
        }
    });
    const playBeepRef = useRef<() => void | null>(null)
    const worker = TimerWorkerManager.getInstance()

    useEffect(() => {
        worker.onmessage(e => {
            const countDownSeconds = e.data

            if (countDownSeconds < 1) {
                if (playBeepRef.current) {
                    playBeepRef.current()
                    playBeepRef.current = null
                }

                dispatch({
                    type: TaskActionsTypes.COMPLETE_TASK
                })
                worker.terminate()
            } else {
                dispatch({
                    type: TaskActionsTypes.COUNT_DOWN,
                    payload: { secondsRemaining: countDownSeconds }
                })
            }
        })
    },)

    useEffect(() => {
        localStorage.setItem('state', JSON.stringify(state))
        if (!state.activeTask) {
            worker.terminate()
        }

        document.title = `${state.formattedSecondsRemaining} - Chronos Pomodoro`

        worker.postMessage(state)
    }, [worker, state])

    useEffect(() => {
        if (state.activeTask && playBeepRef.current === null) {
            playBeepRef.current = loadBeep()
        } else {
            playBeepRef.current = null
        }
    }, [state.activeTask])

    return (
        <TaskContext.Provider value={{ state, dispatch }}>
            {children}
        </TaskContext.Provider>
    );
}